% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individualisation.R
\name{individualisation}
\alias{individualisation}
\title{Individualisation Function}
\usage{
individualisation(
  db = db,
  unknown = unknown,
  Th = 0,
  allele_freq = T,
  known_alleles = T
)
}
\arguments{
\item{db}{Reference population data. File containing genotypes or allele frequencies
for reference population.}

\item{unknown}{Trace profile. File containing genotypes for the case individual.}

\item{Th}{Theta. Value for theta parameter.}

\item{allele_freq}{Reference population data format. Does your reference population
data contain allele frequencies or genotypes? Defaults to allele_freq = T}

\item{known_alleles}{Are the alleles for your case individual equivalent to those in
the reference data? Defaults to known_alleles = T}
}
\description{
A function for calculating RMP for individualisation test
}
\examples{

# load data
data(BadgerRefPop)
data(BadgerTrace)

\dontrun{
individualisation(db = BadgerRefPop,
                  unknown = BadgerTrace,
                  Th = 0.1,
                  allele_freq = T,
                  known_alleles = T)
}

}
\author{
Emily Humble
Kyle Ewart
}
\keyword{individualisation}
